refactor: extract highlighter persistent listeners
